@@ -85,7 +85,7 @@ public class DataMigrationManager {
             List<DataMigration> requiredMigrations = this.migrations.stream()
                     .filter(x -> x.getRevision() > finalCurrentMigration)
                     .sorted(Comparator.comparingInt(DataMigration::getRevision))
-                    .collect(Collectors.toList());
+                    .toList();
 
             // Nothing to migrate, abort
             if (requiredMigrations.isEmpty()) {
